Extract _parse_invocation_response to shared proto.py module
Deduplicate the identical parser from cep.py and pccs.py into proto.py
so both InvocationService clients share a single implementation.
