+++ tests/test_pccs.py
@@ -24,10 +24,10 @@ from custom_components.polestar_soc.pccs import (
     _build_unlock_request,
     _lock_error_context,
     _parse_charge_timer_response,
-    _parse_invocation_response,
     _parse_set_charge_timer_response,
     _parse_target_soc_response,
 )
+from custom_components.polestar_soc.proto import _parse_invocation_response
 from custom_components.polestar_soc.proto import (
     _decode_message,
     _decode_varint,
